Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240119/합과 평균/sum-and-avg.cpp
@@ -3,6 +3,6 @@ using namespace std;
 int main() {
     int a, b;
     cin >> a >> b;
-    cout << a + b << " " << (double)(a + b) / 2;
+    cout << a + b << " " << double(double(a + b) / 2);
     return 0;
 }
